+++ .github/workflows/lint.yml
@@ -13,7 +13,7 @@ jobs:
         uses: actions/checkout@v3
 
       - name: Install
-        run: npm install --legacy-peer-deps
+        run: npm install
 
       - name: Test Lint
         run: npm run lint:fix
