+++ src/components/renders/filters/renderColors.ts
@@ -1,15 +1,12 @@
 import { Product } from '../../../types';
 
 export function renderColors(data: Array<Product>) {
-  const colors: string[] = [];
-
-  data.map((el) => colors.push(el.color));
-
-  const result = colors.filter((e, i) => colors.indexOf(e) === i);
+  const colors = data.map((el) => el.color);
+  const noRepeat = new Set(colors);
 
   const element = document.getElementById('items_color');
 
-  element.innerHTML = result
+  element.innerHTML = Array.from(noRepeat)
     .map((el) => {
       return `<div class="item item-color">
                 <label><input type="checkbox" name="color" id="color" data-color="${el}"/>&#160;${el}</label>
