+++ src/index.ts
@@ -1,7 +1,7 @@
 import './assets/style/style';
 import './modules/actions';
 
-import { renderFilters } from './components/renders/renderFilterss';
+import { renderFilters } from './components/renders/renderFilters';
 import data from './data/db';
 
 renderFilters(data);
